Refactor and add documentation for form element node model
