+++ libs/silo/ngx-lib/src/form-builder/components/form-element-definition-form-portal/form-element-definition-form-portal.component.html
@@ -1,6 +1,6 @@
 <div
   class="silo-form-element-definition-form-portal"
-  [class.active]="nodeModel.state.isActive"
+  [class.active]="nodeModel.instance.isActive"
 >
   <ng-template cdkPortalOutlet></ng-template>
 </div>
