+++ libs/silo/ngx-lib/src/form-builder/form-builder.public-api.ts
@@ -7,8 +7,8 @@ export * from './models/definitions/form-multi-select-definition-model';
 export * from './models/definitions/form-text-definition-model';
 export * from './models/form-builder-events';
 export * from './models/form-definition-types';
+export * from './models/form-element-instance-model';
 export * from './models/form-element-member-model';
 export * from './models/form-element-node-model';
 export * from './models/form-element-registry-config-model';
-export * from './models/form-element-state-model';
 export * from './services/form-builder.service';
